refactor: 코멘트 구조 변경 util 분리
